Removed extra cell file
