--- arcgis-ios-sdk-samples/Layers/RGB renderer/RGB Renderer Cells.swift
@@ -1,28 +1,0 @@
-//
-// Copyright 2016 Esri.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import UIKit
-
-class RGBRendererInputCell: UITableViewCell {
-    @IBOutlet weak var leadingLabel: UILabel!
-    @IBOutlet weak var textField: UITextField!
-}
-
-class RGBRenderer3InputCell: UITableViewCell {
-    @IBOutlet weak var leadingLabel: UILabel!
-    @IBOutlet weak var textField1: UITextField!
-    @IBOutlet weak var textField2: UITextField!
-    @IBOutlet weak var textField3: UITextField!
-}
